+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # nocturn changelog
 
+## 1.1.0 (in dev)
+
+### Filtering
+
+- Added a function "set_min_sleep_period" to keep sessions with a sleep period longer than a threshold (in hours)
+- Added a filter in the app "sleep" tab: "Minimum Time Asleep"
+
+### Data cleaning
+
+- If `session_start` and `sleep_onset_latency` are available, they will be used to infer `time_at_sleep_onset` (if not already provided)
+
 ## 1.0.2 (18/11/2025)
 
 ### Sleep report